Keep the last digit from pulsing on the second.
Now that fetching the time is driven by interrupt, the time taken to fetch the
time would leave the last digit in the cycle enabled for a fractionally longer
amount of time.  This causes the last digit to slightly visibly pulse brighter
when fetching the time.

Disabling the last digit right before fetching the time fixes the problem.